yamanekoの初期描画位置を何とかしたいWIP
